Refactor monolithic components: CalendarView, CreateSessionDialog, Chatbot, Navbar, Room
