+++ src/components/CreateSession/SessionForm.tsx
@@ -1,0 +1,196 @@
+import React from "react";
+import { format } from "date-fns";
+import { Calendar as CalendarIcon, Loader2 } from "lucide-react";
+import { cn } from "@/lib/utils";
+import { UseFormReturn } from "react-hook-form";
+import { FormValues } from "@/hooks/useCreateSession";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { Textarea } from "@/components/ui/textarea";
+import { Calendar } from "@/components/ui/calendar";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
+import { Button } from "@/components/ui/button";
+import { DurationSelector } from "./DurationSelector";
+
+interface SessionFormProps {
+  form: UseFormReturn<FormValues>;
+  onSubmit: (values: FormValues) => Promise<void>;
+  isLoading: boolean;
+  selectedPreset: number;
+  setSelectedPreset: (val: number) => void;
+  useCustom: boolean;
+  setUseCustom: (val: boolean) => void;
+}
+
+export const SessionForm = React.memo(function SessionForm({
+  form,
+  onSubmit,
+  isLoading,
+  selectedPreset,
+  setSelectedPreset,
+  useCustom,
+  setUseCustom,
+}: SessionFormProps) {
+  const inputCls =
+    "bg-white/5 border-white/10 text-white placeholder:text-gray-500 focus-visible:ring-cyan-500";
+
+  return (
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4 mt-4">
+        {/* Title */}
+        <FormField
+          control={form.control}
+          name="title"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Title</FormLabel>
+              <FormControl>
+                <Input
+                  placeholder="e.g. React Hooks Deep Dive"
+                  className={inputCls}
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage className="text-red-400" />
+            </FormItem>
+          )}
+        />
+
+        {/* Description */}
+        <FormField
+          control={form.control}
+          name="description"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Description</FormLabel>
+              <FormControl>
+                <Textarea
+                  placeholder="What will you cover in this session?"
+                  className={`${inputCls} resize-none`}
+                  rows={3}
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage className="text-red-400" />
+            </FormItem>
+          )}
+        />
+
+        {/* Date + Time row */}
+        <div className="grid grid-cols-2 gap-4">
+          <FormField
+            control={form.control}
+            name="date"
+            render={({ field }) => (
+              <FormItem className="flex flex-col">
+                <FormLabel>Date</FormLabel>
+                <Popover>
+                  <PopoverTrigger asChild>
+                    <FormControl>
+                      <Button
+                        type="button"
+                        variant="outline"
+                        className={cn(
+                          "w-full pl-3 text-left font-normal bg-white/5 border-white/10 text-white hover:bg-white/10 hover:text-white",
+                          !field.value && "text-gray-500"
+                        )}
+                      >
+                        {field.value ? (
+                          format(field.value, "PPP")
+                        ) : (
+                          <span>Pick a date</span>
+                        )}
+                        <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
+                      </Button>
+                    </FormControl>
+                  </PopoverTrigger>
+                  <PopoverContent
+                    className="w-auto p-0 bg-[#0f172a] border-white/10"
+                    align="start"
+                  >
+                    <Calendar
+                      mode="single"
+                      selected={field.value}
+                      onSelect={field.onChange}
+                      disabled={(date) =>
+                        date < new Date(new Date().setHours(0, 0, 0, 0))
+                      }
+                      initialFocus
+                      className="text-white"
+                    />
+                  </PopoverContent>
+                </Popover>
+                <FormMessage className="text-red-400" />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="time"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Time</FormLabel>
+                <FormControl>
+                  <Input type="time" className={inputCls} {...field} />
+                </FormControl>
+                <FormMessage className="text-red-400" />
+              </FormItem>
+            )}
+          />
+        </div>
+
+        {/* Duration selector */}
+        <DurationSelector
+          form={form}
+          selectedPreset={selectedPreset}
+          setSelectedPreset={setSelectedPreset}
+          useCustom={useCustom}
+          setUseCustom={setUseCustom}
+          inputCls={inputCls}
+        />
+
+        {/* Seat Limit */}
+        <FormField
+          control={form.control}
+          name="seatLimit"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Seat Limit (Optional)</FormLabel>
+              <FormControl>
+                <Input
+                  type="number"
+                  min={1}
+                  placeholder="e.g. 50"
+                  className={inputCls}
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage className="text-red-400" />
+            </FormItem>
+          )}
+        />
+
+        <Button
+          type="submit"
+          className="w-full bg-gradient-to-r from-cyan-400 to-purple-500 text-black font-bold hover:opacity-90 transition mt-2"
+          disabled={isLoading}
+        >
+          {isLoading && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
+          {isLoading ? "Scheduling..." : "Schedule Session"}
+        </Button>
+      </form>
+    </Form>
+  );
+});
